+++ .github/workflows/test.yaml
@@ -30,7 +30,7 @@ jobs:
         CONDA_PREFIX: /usr/share/miniconda
       run: |
         pip install pytest pytest-cov
-        pytest tests --cov=lc_macro_pipeline --cov-report=xml
+        pytest tests --cov=laserfarm --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1
       with:
