+++ README.md
@@ -1,10 +1,10 @@
 # Laserchicken Macro(-ecology) Pipeline
 
-[![Actions Status](https://github.com/eEcoLiDAR/lcMacroPipeline/workflows/build%20and%20test/badge.svg?branch=development)](https://github.com/eEcoLiDAR/lcMacroPipeline/actions)
-[![codecov](https://codecov.io/gh/eEcoLiDAR/lcMacroPipeline/branch/development/graph/badge.svg)](http://codecov.io/github/eEcoLiDAR/lcMacroPipeline/branch/development)
-[![Documentation Status](https://readthedocs.org/projects/lcmacropipeline/badge/?version=latest)](https://lcmacropipeline.readthedocs.io/en/latest/?badge=latest)
+[![Actions Status](https://github.com/eEcoLiDAR/Laserfarm/workflows/build%20and%20test/badge.svg?branch=development)](https://github.com/eEcoLiDAR/Laserfarm/actions)
+[![codecov](https://codecov.io/gh/eEcoLiDAR/Laserfarm/branch/development/graph/badge.svg)](http://codecov.io/github/eEcoLiDAR/Laserfarm/branch/development)
+[![Documentation Status](https://readthedocs.org/projects/Laserfarm/badge/?version=latest)](https://laserfarm.readthedocs.io/en/latest/?badge=latest)
 
-lcMacroPipeline provides a FOSS wrapper to [Laserchicken](https://github.com/eEcoLiDAR/laserchicken) supporting the use of
+Laserfarm provides a FOSS wrapper to [Laserchicken](https://github.com/eEcoLiDAR/laserchicken) supporting the use of
 massive LiDAR point cloud data sets for macro-ecology, from data preparation to scheduling and execution
 of distributed processing across a cluster of compute nodes.
 
@@ -12,14 +12,14 @@ of distributed processing across a cluster of compute nodes.
 
 The package can be downloaded using `git`:
 ```shell script
-git clone git@github.com:eEcoLiDAR/lcMacroPipeline.git
+git clone git@github.com:eEcoLiDAR/Laserfarm.git
 ```
 It requires the [PDAL](https://pdal.io) and [GDAL](https://gdal.org) libraries and the PDAL Python 
 bindings. These packages are most easily installed through `conda` from the `conda-forge` channel. The 
 remaining dependencies can be retrieved and installed using `pip`:
 ```shell script
 conda install pdal python-pdal gdal -c conda-forge
-cd lcMacroPipeline && pip install . 
+cd Laserfarm && pip install . 
 ```
 Alternatively, a new environment with the package and all its dependencies can be created from the
 YAML file provided:
@@ -29,7 +29,7 @@ conda env create -f environment.yml
 
 # Documentation
 
-The project's full documentation can be found [here](https://lcmacropipeline.readthedocs.io/en/latest/).
+The project's full documentation can be found [here](https://laserfarm.readthedocs.io/en/latest/).
 
 # Applications and Current Limitations
 
@@ -40,7 +40,7 @@ as well, this has not been tested and it is generally expected to fail.
 
 # Contributing
 
-If you want to contribute to the development of lcMacroPipeline,
+If you want to contribute to the development of Laserfarm,
 have a look at the  [contribution guidelines](CONTRIBUTING.md).
 
 # License
