+++ docs/index.rst
@@ -1,23 +1,23 @@
-.. lc_macro_pipeline documentation master file, created by
+.. laserfarm documentation master file, created by
    sphinx-quickstart on Thu May  7 23:16:29 2020.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to lc_macro_pipeline's documentation!
+Welcome to laserfarm's documentation!
 =============================================
 
-lcMacroPipeline provides a FOSS wrapper to Laserchicken supporting the use of massive LiDAR point cloud
+Laserfarm provides a FOSS wrapper to Laserchicken supporting the use of massive LiDAR point cloud
 data sets for macro-ecology, from data preparation to scheduling and execution of distributed processing
 across a cluster of compute nodes.
 
-lcMacroPipeline has been developed within the `eEcoLiDAR project`_ at `the Netherlands eScience Center`_.
-The code is hosted on `GitHub`_. If you want to contribute to the development of lcMacroPipeline,
+Laserfarm has been developed within the `eEcoLiDAR project`_ at `the Netherlands eScience Center`_.
+The code is hosted on `GitHub`_. If you want to contribute to the development of Laserfarm,
 have a look at the  `contribution guidelines`_.
 
 .. _eEcoLiDAR project: https://www.esciencecenter.nl/projects/eecolidar/
 .. _the Netherlands eScience Center: https://www.esciencecenter.nl
-.. _GitHub: https://github.com/eEcoLiDAR/lcMacroPipeline
-.. _contribution guidelines: https://github.com/eEcoLiDAR/lcMacroPipeline/tree/master/CONTRIBUTING.md
+.. _GitHub: https://github.com/eEcoLiDAR/Laserfarm
+.. _contribution guidelines: https://github.com/eEcoLiDAR/Laserfarm/tree/master/CONTRIBUTING.md
 
 .. toctree::
    :maxdepth: 2
