+++ laserfarm/__init__.py
@@ -1,0 +1,12 @@
+__license__ = 'Apache Licence 2.0'
+__author__ = 'Netherlands eScience Center'
+__email__ = 'team-atlas@esciencecenter.nl'
+
+from .__version__ import __version__
+
+from laserfarm.data_processing import DataProcessing
+from laserfarm.geotiff_writer import GeotiffWriter
+from laserfarm.retiler import Retiler
+from laserfarm.classification import Classification
+
+from laserfarm.macro_pipeline import MacroPipeline
