--- lc_macro_pipeline/__init__.py
@@ -1,12 +1,0 @@
-__license__ = 'Apache Licence 2.0'
-__author__ = 'Netherlands eScience Center'
-__email__ = 'team-atlas@esciencecenter.nl'
-
-from .__version__ import __version__
-
-from lc_macro_pipeline.data_processing import DataProcessing
-from lc_macro_pipeline.geotiff_writer import GeotiffWriter
-from lc_macro_pipeline.retiler import Retiler
-from lc_macro_pipeline.classification import Classification
-
-from lc_macro_pipeline.macro_pipeline import MacroPipeline
